Consolidation: test_caiso.py
